SERVER-6400: cleaning api for performance
Conflicts:

	src/mongo/db/matcher/expression_array.cpp
	src/mongo/db/matcher/expression_leaf.cpp
	src/mongo/db/matcher/matchable.cpp
	src/mongo/db/matcher/matchable.h
	src/mongo/db/matcher/matcher.cpp

+++ src/mongo/db/matcher/expression_internal.cpp
@@ -30,7 +30,7 @@ namespace mongo {
 
     BSONElement getFieldDottedOrArray( const BSONObj& doc,
                                        const FieldRef& path,
-                                       int32_t* idxPath,
+                                       size_t* idxPath,
                                        bool* inArray )  {
         BSONElement res;
